# openapi.yaml
openapi: 3.0.1
info:
  title: Laine Déco API
  version: 1.0.0
  description: |
    Documentation complète de l'API Laine Déco - Plateforme de vente de laine et accessoires

    ## 🔐 Authentification
    Utilisez un token JWT Bearer dans l'en-tête `Authorization: Bearer <token>`

    ## 📋 Sections de l'API
    - **Public** : Routes accessibles sans authentification
    - **Client** : Routes pour les clients authentifiés
    - **Admin** : Routes pour les administrateurs
    - **Finance** : Routes pour l'équipe finance

  contact:
    name: Support Laine Déco
    email: support@laine-deco.cm
    url: https://laine-deco.cm
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Serveur local
  - url: https://api.laine-deco.cm
    description: Serveur de production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Entrez le token JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/client/notifications/{id}:
    get:
      tags:
        - Client - Notifications
      summary: Détails notification
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification
  /api/client/notifications/{id}/mark-as-read:
    post:
      tags:
        - Client - Notifications
      summary: Marquer comme lu
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Marqué comme lu
...
  /api/finance/expenses:
    post:
      tags:
        - Finance - Expenses
      summary: Créer une dépense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Dépense créée

    get:
      tags:
        - Finance - Expenses
      summary: Lister les dépenses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dépenses
  /api/finance/expenses/by-status:
    get:
      tags:
        - Finance - Expenses
      summary: Dépenses par statut
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dépenses
  /api/finance/expenses/by-category:
    get:
      tags:
        - Finance - Expenses
      summary: Dépenses par catégorie
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dépenses
  /api/finance/expenses/{id}/status:
    put:
      tags:
        - Finance - Expenses
      summary: Mettre à jour le statut
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Statut mis à jour
  /api/finance/expenses/statistics/total-by-status:
    get:
      tags:
        - Finance - Expenses
      summary: Total des dépenses par statut
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Total
  /api/finance/expenses/statistics/total-by-category:
    get:
      tags:
        - Finance - Expenses
      summary: Total des dépenses par catégorie
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Total
...